Feature/pagination

+++ domain/src/main/java/com/app/domain/models/interestingness/response/Photos.kt
@@ -1,9 +1,9 @@
 package com.app.domain.models.interestingness.response
 
 data class Photos(
-    val page: Long,
-    val pages: Long,
-    val perpage: Long,
-    val total: Long,
+    val page: Int,
+    val pages: Int,
+    val perpage: Int,
+    val total: Int,
     val photo: List<Photo>
 )
